+++ src/main/java/ureca/muneobe/common/slang/SlangController.java
@@ -1,4 +1,4 @@
-package ureca.muneobe.slang;
+package ureca.muneobe.common.slang;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -9,11 +9,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ureca.muneobe.global.response.ErrorCode;
 import ureca.muneobe.global.response.ResponseBody;
-import ureca.muneobe.slang.dto.AddSlangRequest;
-import ureca.muneobe.slang.dto.AddSlangResponse;
-import ureca.muneobe.slang.dto.DeleteSlangRequest;
-import ureca.muneobe.slang.dto.DeleteSlangResponse;
-import ureca.muneobe.slang.service.SlangService;
+import ureca.muneobe.common.slang.dto.AddSlangRequest;
+import ureca.muneobe.common.slang.dto.AddSlangResponse;
+import ureca.muneobe.common.slang.dto.DeleteSlangRequest;
+import ureca.muneobe.common.slang.dto.DeleteSlangResponse;
+import ureca.muneobe.common.slang.service.SlangService;
 
 import java.util.List;
