+++ src/main/java/ureca/muneobe/common/slang/dto/AddSlangRequest.java
@@ -1,4 +1,4 @@
-package ureca.muneobe.slang.dto;
+package ureca.muneobe.common.slang.dto;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Getter
 @Setter
-public class DeleteSlangRequest {
+public class AddSlangRequest {
 
     private List<String> words;
 }
