+++ src/main/java/ureca/muneobe/common/slang/service/AhoCorasickSlangFilterService.java
@@ -1,4 +1,4 @@
-package ureca.muneobe.slang.service;
+package ureca.muneobe.common.slang.service;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
