+++ src/main/java/ureca/muneobe/common/slang/service/ModerationApiSlangFilterService.java
@@ -1,11 +1,10 @@
-package ureca.muneobe.slang.service;
+package ureca.muneobe.common.slang.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
 
 import java.io.IOException;
 import java.net.URI;
