+++ src/test/java/ureca/muneobe/slang/SlangFilterServiceTest.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationContext;
 import org.springframework.test.context.ActiveProfiles;
-import ureca.muneobe.slang.service.SlangFilterService;
+import ureca.muneobe.common.slang.service.SlangFilterService;
 
 import java.time.Duration;
 import java.time.Instant;
